powerhint: Add nodes to control per-task iowait boost
Port from 27901981

It's disabled by default.

Ensure camera and game mode explicitly disable it too as it is only
required when auto margins are enabled for now.

Ensure it is disabled when screen is off too as We don't want io
activities to cause wake up on bigger cores while screen is off.

Bug: 353256081
Test: Local Genshin test
Flag: NONE powerhint change does not support flags

Change-Id: Ic4ec943181885be2a788c2b8754e814826abdd24